requirements/devel.txt
-r common.txt

ipython==0.13.1
django-debug-toolbar==0.9.4
